refactor(arch): Split applications router into borrower_documents, emails, downloads routers, closes #200
