+++ app/routers/applications.py
@@ -1,16 +1,10 @@
-import io
 import logging
-import zipfile
 from datetime import datetime
 from typing import List
 
-import pandas as pd
 from botocore.exceptions import ClientError
-from fastapi import APIRouter, BackgroundTasks, Depends, Form, HTTPException, Query, UploadFile, status
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, status
 from fastapi.encoders import jsonable_encoder
-from fastapi.responses import StreamingResponse
-from reportlab.lib.pagesizes import letter
-from reportlab.platypus import Paragraph, SimpleDocTemplate, Spacer
 from sqlalchemy import and_, text
 from sqlalchemy.orm import Session, joinedload
 
@@ -19,11 +13,9 @@ from app import models, parsers, serializers
 from app.auth import OCP_only, get_current_user, get_user
 from app.aws import CognitoClient, get_cognito_client
 from app.db import get_db, transaction_session
-from app.i18n import get_translated_string
 from app.settings import app_settings
-from app.utils import background, tables
+from app.utils import background
 from app.utils.statistics import update_statistics
-from reportlab_mods import styleSubTitle, styleTitle
 
 logger = logging.getLogger(__name__)
 
@@ -238,323 +230,6 @@ async def approve_application(
 
 
 @router.post(
-    "/applications/change-email",
-    tags=["applications"],
-    response_model=parsers.ChangeEmail,
-)
-async def change_email(
-    payload: parsers.ChangeEmail,
-    session: Session = Depends(get_db),
-    client: CognitoClient = Depends(get_cognito_client),
-):
-    """
-    Change the email address for an application.
-
-    :param payload: The data for changing the email address.
-    :type payload: parsers.ChangeEmail
-
-    :param session: The database session.
-    :type session: Session
-
-    :param client: The Cognito client.
-    :type client: CognitoClient
-
-    :return: The data for changing the email address.
-    :rtype: parsers.ChangeEmail
-
-    """
-    with transaction_session(session):
-        application = utils.get_application_by_uuid(payload.uuid, session)
-        old_email = application.primary_email
-        confirmation_email_token = utils.update_application_primary_email(application, payload.new_email)
-        models.ApplicationAction.create(
-            session,
-            type=models.ApplicationActionType.MSME_CHANGE_EMAIL,
-            data=jsonable_encoder(payload, exclude_unset=True),
-            application_id=application.id,
-        )
-
-        external_message_id = client.send_new_email_confirmation_to_sme(
-            application.borrower.legal_name,
-            payload.new_email,
-            old_email,
-            confirmation_email_token,
-            application.uuid,
-        )
-
-        models.Message.create(
-            session,
-            application=application,
-            type=models.MessageType.EMAIL_CHANGE_CONFIRMATION,
-            external_message_id=external_message_id,
-        )
-
-        return payload
-
-
-@router.post(
-    "/applications/confirm-change-email",
-    tags=["applications"],
-    response_model=parsers.ChangeEmail,
-)
-async def confirm_email(
-    payload: parsers.ConfirmNewEmail,
-    session: Session = Depends(get_db),
-):
-    """
-    Confirm the email address change for an application.
-
-    :param payload: The data for confirming the email address change.
-    :type payload: parsers.ConfirmNewEmail
-
-    :param session: The database session.
-    :type session: Session
-
-    :return: The data for confirming the email address change.
-    :rtype: parsers.ChangeEmail
-
-    """
-    with transaction_session(session):
-        application = utils.get_application_by_uuid(payload.uuid, session)
-
-        utils.check_pending_email_confirmation(application, payload.confirmation_email_token)
-
-        models.ApplicationAction.create(
-            session,
-            type=models.ApplicationActionType.MSME_CONFIRM_EMAIL,
-            data=jsonable_encoder(payload, exclude_unset=True),
-            application_id=application.id,
-        )
-
-        return parsers.ChangeEmail(new_email=application.primary_email, uuid=application.uuid)
-
-
-@router.get(
-    "/applications/documents/id/{id}",
-    tags=["applications"],
-)
-async def get_borrower_document(
-    id: int,
-    session: Session = Depends(get_db),
-    user: models.User = Depends(get_user),
-):
-    """
-    Retrieve a borrower document by its ID and stream the file content as a response.
-
-    :param id: The ID of the borrower document to retrieve.
-    :type id: int
-
-    :param session: The database session.
-    :type session: Session
-
-    :param user: The current user.
-    :type user: models.User
-
-    :return: A streaming response with the borrower document file content.
-    :rtype: StreamingResponse
-
-    """
-    with transaction_session(session):
-        document = models.BorrowerDocument.first_by(session, "id", id)
-        utils.get_file(document, user, session)
-
-        def file_generator():
-            yield document.file
-
-        headers = {
-            "Content-Disposition": f'attachment; filename="{document.name}"',
-            "Content-Type": "application/octet-stream",
-        }
-        return StreamingResponse(file_generator(), headers=headers)
-
-
-@router.get(
-    "/applications/{application_id}/download-application/{lang}",
-    tags=["applications"],
-)
-async def download_application(
-    application_id: int,
-    lang: str,
-    session: Session = Depends(get_db),
-    user: models.User = Depends(get_user),
-):
-    """
-    Retrieve all documents related to an application and stream them as a zip file.
-
-    :param application_id: The ID of the application to retrieve documents for.
-    :type application_id: int
-
-    :param session: The database session.
-    :type session: Session
-
-    :param user: The current user.
-    :type user: models.User
-
-    :return: A streaming response with a zip file containing the documents.
-    :rtype: StreamingResponse
-    """
-    with transaction_session(session):
-        application = utils.get_application_by_id(application_id, session)
-
-        borrower = application.borrower
-        award = application.award
-
-        documents = (
-            session.query(models.BorrowerDocument)
-            .filter(models.BorrowerDocument.application_id == application_id)
-            .all()
-        )
-
-        previous_awards = utils.get_previous_awards(application, session)
-
-        buffer = io.BytesIO()
-        doc = SimpleDocTemplate(buffer, pagesize=letter)
-
-        elements = []
-
-        elements.append(Paragraph(get_translated_string("Application Details", lang), styleTitle))
-
-        elements.append(tables.create_application_table(application, lang))
-        elements.append(Spacer(1, 20))
-        elements.append(tables.create_borrower_table(borrower, application, lang))
-        elements.append(Spacer(1, 20))
-        elements.append(tables.create_documents_table(documents, lang))
-        elements.append(Spacer(1, 20))
-        elements.append(tables.create_award_table(award, lang))
-
-        if previous_awards and len(previous_awards) > 0:
-            elements.append(Spacer(1, 20))
-            elements.append(Paragraph(get_translated_string("Previous Public Sector Contracts", lang), styleSubTitle))
-            for award in previous_awards:
-                elements.append(tables.create_award_table(award, lang))
-                elements.append(Spacer(1, 20))
-
-        doc.build(elements)
-
-        name = get_translated_string("Application Details", lang).replace(" ", "_")
-        filename = f"{name}-{application.borrower.legal_identifier}" + f"-{application.award.source_contract_id}.pdf"
-
-        in_memory_zip = io.BytesIO()
-        with zipfile.ZipFile(in_memory_zip, "w") as zip_file:
-            zip_file.writestr(filename, buffer.getvalue())
-            for document in documents:
-                zip_file.writestr(document.name, document.file)
-
-        application_action_type = (
-            models.ApplicationActionType.OCP_DOWNLOAD_APPLICATION
-            if user.is_OCP()
-            else models.ApplicationActionType.FI_DOWNLOAD_APPLICATION
-        )
-        models.ApplicationAction.create(
-            session,
-            type=application_action_type,
-            application_id=application.id,
-            user_id=user.id,
-        )
-
-        headers = {
-            "Content-Disposition": f'attachment; filename="{filename}"',
-            "Content-Type": "application/zip",
-        }
-
-        return StreamingResponse(io.BytesIO(in_memory_zip.getvalue()), headers=headers)
-
-
-@router.post(
-    "/applications/upload-document",
-    tags=["applications"],
-    response_model=models.BorrowerDocumentBase,
-)
-async def upload_document(
-    file: UploadFile,
-    uuid: str = Form(...),
-    type: str = Form(...),
-    session: Session = Depends(get_db),
-):
-    """
-    Upload a document for an application.
-
-    :param file: The uploaded file.
-    :type file: UploadFile
-
-    :param uuid: The UUID of the application.
-    :type uuid: str
-
-    :param type: The type of the document.
-    :type type: str
-
-    :param session: The database session.
-    :type session: Session
-
-    :return: The created or updated borrower document.
-    :rtype: models.BorrowerDocumentBase
-
-    """
-    with transaction_session(session):
-        new_file, filename = utils.validate_file(file)
-        application = utils.get_application_by_uuid(uuid, session)
-        if not application.pending_documents:
-            raise HTTPException(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                detail="Cannot upload document at this stage",
-            )
-
-        document = utils.create_or_update_borrower_document(filename, application, type, session, new_file)
-
-        models.ApplicationAction.create(
-            session,
-            type=models.ApplicationActionType.MSME_UPLOAD_DOCUMENT,
-            data={"file_name": filename},
-            application_id=application.id,
-        )
-
-        return document
-
-
-@router.post(
-    "/applications/upload-contract",
-    tags=["applications"],
-    response_model=models.BorrowerDocumentBase,
-)
-async def upload_contract(
-    file: UploadFile,
-    uuid: str = Form(...),
-    session: Session = Depends(get_db),
-):
-    """
-    Upload a contract document for an application.
-
-    :param file: The uploaded file.
-    :type file: UploadFile
-
-    :param uuid: The UUID of the application.
-    :type uuid: str
-
-    :param session: The database session.
-    :type session: Session
-
-    :return: The created or updated borrower document representing the contract.
-    :rtype: models.BorrowerDocumentBase
-
-    """
-    with transaction_session(session):
-        new_file, filename = utils.validate_file(file)
-        application = utils.get_application_by_uuid(uuid, session)
-
-        utils.check_application_status(application, models.ApplicationStatus.APPROVED)
-
-        document = utils.create_or_update_borrower_document(
-            filename,
-            application,
-            models.BorrowerDocumentType.SIGNED_CONTRACT,
-            session,
-            new_file,
-        )
-
-        return document
-
-
-@router.post(
     "/applications/confirm-upload-contract",
     tags=["applications"],
     response_model=serializers.ApplicationResponse,
@@ -624,61 +299,6 @@ async def confirm_upload_contract(
             documents=application.borrower_documents,
             creditProduct=application.credit_product,
         )
-
-
-@router.post(
-    "/applications/{id}/upload-compliance",
-    tags=["applications"],
-    response_model=models.BorrowerDocumentBase,
-)
-async def upload_compliance(
-    id: int,
-    file: UploadFile,
-    session: Session = Depends(get_db),
-    user: models.User = Depends(get_user),
-):
-    """
-    Upload a compliance document for an application.
-
-    :param id: The ID of the application.
-    :type id: int
-
-    :param file: The uploaded file.
-    :type file: UploadFile
-
-    :param session: The database session.
-    :type session: Session
-
-    :param user: The current user.
-    :type user: models.User
-
-    :return: The created or updated borrower document representing the compliance report.
-    :rtype: models.BorrowerDocumentBase
-
-    """
-    with transaction_session(session):
-        new_file, filename = utils.validate_file(file)
-        application = utils.get_application_by_id(id, session)
-
-        utils.check_FI_user_permission(application, user)
-
-        document = utils.create_or_update_borrower_document(
-            filename,
-            application,
-            models.BorrowerDocumentType.COMPLIANCE_REPORT,
-            session,
-            new_file,
-            True,
-        )
-
-        models.ApplicationAction.create(
-            session,
-            type=models.ApplicationActionType.FI_UPLOAD_COMPLIANCE,
-            data={"file_name": filename},
-            application_id=application.id,
-        )
-
-        return document
 
 
 @router.put(
@@ -1009,24 +629,6 @@ async def start_application(
         application.lender_started_at = datetime.now(application.created_at.tzinfo)
         background_tasks.add_task(update_statistics)
         return application
-
-
-@router.get(
-    "/applications/export/{lang}",
-    tags=["applications"],
-    response_class=StreamingResponse,
-)
-async def export_applications(
-    lang: str,
-    user: models.User = Depends(get_user),
-    session: Session = Depends(get_db),
-):
-    df = pd.DataFrame(utils.get_all_fi_applications_emails(session, user.lender_id, lang))
-    stream = io.StringIO()
-    df.to_csv(stream, index=False)
-    response = StreamingResponse(iter([stream.getvalue()]), media_type="text/csv")
-    response.headers["Content-Disposition"] = "attachment; filename=export.csv"
-    return response
 
 
 @router.get(
